+++ MqMessaging/MqChannel.cpp
@@ -304,7 +304,7 @@ void MqChannel::sendTo(const std::basic_string<unsigned char>& message)
 {
   TRC_INFORMATION("Send to MQ: " << std::endl << MEM_HEX(message.data(), message.size()));
 
-  unsigned long toWrite = message.size();
+  unsigned long toWrite = static_cast<unsigned long>(message.size());
   unsigned long written = 0;
   bool reconnect = false;
   bool fSuccess;
